# Conflicts:
#	third_party/libpng/ci/ci_autotools.sh
#	third_party/libpng/ci/ci_cmake.sh
#	third_party/libpng/ci/ci_legacy.sh
#	third_party/libpng/scripts/makefile.vcwin-arm64

+++ filament/src/Froxelizer.cpp
@@ -105,6 +105,10 @@ Froxelizer::Froxelizer(FEngine& engine)
           mZLightNear(FROXEL_FIRST_SLICE_DEPTH),
           mZLightFar(FROXEL_LAST_SLICE_DISTANCE)
 {
+    if (UTILS_UNLIKELY(engine.getActiveFeatureLevel() == FeatureLevel::FEATURE_LEVEL_0)) {
+        return;
+    }
+
     DriverApi& driverApi = engine.getDriverApi();
 
     static_assert(std::is_same_v<RecordBufferType, uint8_t>,
@@ -132,8 +136,12 @@ void Froxelizer::terminate(DriverApi& driverApi) noexcept {
     mPlanesX = nullptr;
     mDistancesZ = nullptr;
 
-    driverApi.destroyBufferObject(mRecordsBuffer);
-    driverApi.destroyTexture(mFroxelTexture);
+    if (mRecordsBuffer) {
+        driverApi.destroyBufferObject(mRecordsBuffer);
+    }
+    if (mFroxelTexture) {
+        driverApi.destroyTexture(mFroxelTexture);
+    }
 }
 
 void Froxelizer::setOptions(float zLightNear, float zLightFar) noexcept {
@@ -559,6 +567,8 @@ void Froxelizer::froxelizeLoop(FEngine& engine,
     auto process = [ this, &froxelThreadData,
                      spheres, directions, instances, &viewMatrix, &lcm ]
             (size_t count, size_t offset, size_t stride) {
+
+        SYSTRACE_NAME("FroxelizeLoop Job");
 
         const mat4f& projection = mProjection;
         const mat3f& vn = viewMatrix.upperLeft();
